requirements.txt
biopython==1.76
pandas==1.2.2
levenshtein==0.12.0
jinja2==2.11.3
pyyaml==5.4.1
